arm64/ascend: Implement the read_persistend_clock64 for aarch64
ascend inclusion
category: feature
bugzilla: NA
CVE: NA

-------------------------------------------------

The read_persistend_clock and read_persistend_clock64 is designed to
be implemented at the platform level, which make it impossible to
compile these platform in a single kernel.

Implement the functions at the architecture level, and provide a
registration interface for read_persistend_clock64, the platform
private drivers no need to be compiled in a single kernel.

This patch is like the commit bd0493eaaf5c7 ("ARM: 7413/1: move read...")

Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>