arm64: Kconfig: change fix compile error if gcc don't support armv8.4-a
hulk inclusion
category: feature
CVE: NA

-----------------------

After 4d0831e8a029 ("kconfig: unify cc-option and as-option")
in mainline, we can use cc-option.

But in linux-4.19, this patch is not merged, so we will get the
follow error:
  CC      scripts/mod/empty.o
  CC      scripts/mod/devicetable-offsets.s
Assembler messages:
Error: unknown architecture `armv8.4-a'

Error: unrecognized option -march=armv8.4-a

Fix this by changing cc-option to as-option.

Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>
Reviewed-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>