percpu: update free path with correct new free region
mainline inclusion
from mainline-5.2-rc1
commit 8e5a2b9893f36457582596fdade10f6feb2797ee
category: bugfix
bugzilla: 34611
CVE: NA

-------------------------------------------------
When updating the chunk's contig_hint on the free path of a hint that
does not touch the page boundaries, it was incorrectly using the
starting offset of the free region and the block's contig_hint. This
could lead to incorrect assumptions about fit given a size and better
alignment of the start. Fix this by using (end - start) as this is only
called when updating a hint within a block.

Signed-off-by: Dennis Zhou <dennis@kernel.org>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit 8e5a2b9893f36457582596fdade10f6feb2797ee)
Signed-off-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Liu Shixin <liushixin2@huawei.com>
Reviewed-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>