xen/events: fix race in evtchn_fifo_unmask()
mainline inclusion
from mainline-v5.10
commit f01337197419b7e8a492e83089552b77d3b5fb90
category: bugfix
bugzilla: NA
CVE: CVE-2020-27673

--------------------------------

Unmasking a fifo event channel can result in unmasking it twice, once
directly in the kernel and once via a hypercall in case the event was
pending.

Fix that by doing the local unmask only if the event is not pending.

This is part of XSA-332.

Cc: stable@vger.kernel.org
Signed-off-by: Juergen Gross <jgross@suse.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>
Reviewed-by: Jason Yan <yanaijie@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>