ascend: share_pool: add support proc_sharepool_init and is_vm_huge_special
ascend inclusion
category: feature
bugzilla: NA
CVE: NA

-------------------------------------------------

The proc_share_init should be used in proc_root_init when enable
the share pool features.

The is_vm_huge_special() was used to distinguish the vma with
VM_HUGE_SPECIAL and handle it, it would not be used by default.

Signed-off-by: Tang Yizhou <tangyizhou@huawei.com>
Signed-off-by: Li Ming <limingming.li@huawei.com>
Signed-off-by: Zefan Li <lizefan@huawei.com>
Signed-off-by: Zhou Guanghui <zhouguanghui1@huawei.com>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>