ascend: share_pool: support fork() and exit() to handle the mm
ascend inclusion
category: feature
bugzilla: NA
CVE: NA

-------------------------------------------------

The fork() will create the new mm for new process, the mm should not
take any information from the parent process, so need to clean it.

The exit() will mmput the mm and free the memory, if the mm is alrready
be used for sp_group, need to clean the group first.

Signed-off-by: Tang Yizhou <tangyizhou@huawei.com>
Signed-off-by: Li Ming <limingming.li@huawei.com>
Signed-off-by: Zefan Li <lizefan@huawei.com>
Signed-off-by: Zhou Guanghui <zhouguanghui1@huawei.com>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>
Reviewed-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>