vfio: avoid to call trace in the ARM64 architecture
euleros inclusion
category: bugfix
bugzilla: 46842
CVE: NA

-------------------------------------------------

In the ARM64 architecture, zero physical address can be
used for system ram, so virtual machine can allocate
this segment of memory as vm's memory, but in
vfio_unmap_unpin() if it found memory physical address
is zero, it will call trace. By the way, x86 reserve
0~4k and avoid to happen this scenario.

So avoid to call trace in the ARM64 architecture.

Signed-off-by: Yiwen Jiang <jiangyiwen@huawei.com>
Signed-off-by: Xiangyou Xie <xiexiangyou@huawei.com>
Reviewed-by: Ying Fang <fangying1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>