KVM: arm64: support vtimer interrupt advance injection
euleros inclusion
category: feature
bugzilla: 46842
CVE: NA

-------------------------------------------------

There is a delay when injecting a vtimer interrupt. And When vcpu
wfx vmexit, it will switch to bgtimer, and the interrupt simulation
path will become longer.

Provides a mechanism for bgtimer to trigger in advance, which can
be injected into the interrupt in advance to avoid delay.

Signed-off-by: Xiangyou Xie <xiexiangyou@huawei.com>
Reviewed-by: Ying Fang <fangying1@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>