evm: Reset IMA flags if EVM ignore mode is selected
hulk inclusion
category: feature
feature: digest-lists

---------------------------

The EVM ignore mode works similarly to the metadata modification mode. They
both allow an operation to be performed even if the operation causes
metadata to become invalid.

Currently, evm_reset_status() notifies to IMA that an operation modified
metadata only when the metadata modification mode was chosen. This patch
sends a notification also when the ignore mode is selected.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>